# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserEntity": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"password",
					"createdAt",
					"updatedAt",
					"deletedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateUserDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponsePayload": {
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserCreateRequestDto": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Real Estate Ecommerce",
		"version": "1.0.0",
		"description": "One place for all things",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/auth/register": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserEntity"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserDto"
							}
						}
					}
				}
			}
		},
		"/{teamId}/user": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"deletedAt": {
											"type": "string",
											"format": "date-time",
											"nullable": true
										},
										"updatedAt": {
											"type": "string",
											"format": "date-time"
										},
										"createdAt": {
											"type": "string",
											"format": "date-time"
										},
										"name": {
											"type": "string"
										},
										"password": {
											"type": "string"
										},
										"email": {
											"type": "string"
										},
										"id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"deletedAt",
										"updatedAt",
										"createdAt",
										"name",
										"password",
										"email",
										"id"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "BadRequest",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponsePayload"
								}
							}
						}
					},
					"409": {
						"description": "ConflictException",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponsePayload"
								}
							}
						}
					}
				},
				"summary": "회원가입",
				"tags": [
					"/user"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "teamId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCreateRequestDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:9000/"
		}
	]
}